Update Filter Run Prefix (Examples).tid
This catches all filter run prefixes. This is superior to the current hardcoded text that doesn't give any clues about non-documented prefixes.

My *actual* wish is to implement the more elaborate changes outlined [here](https://talk.tiddlywiki.org/t/ranty-proposal-for-docs-on-filter-run-prefixes/5772)... but this current PR is at least an improvement meanwhile.