In if widget, use real widgets, not pseudo-widgets
This moves most of the logic out of the `if` widget and into the
condition widgets (then/elseif/else). The only thing the `if` widget
is responsible for is setting up the singly-linked list of condition
widgets; they are then responsible for carrying out the short-circuit
logic and refreshing themselves as needed.