Update ViewWidget.tid to include common variants (#3895)
* Update ViewWidget.tid to include common varients

The provision of common variates to the view widget provides new an experience user examples of the different variants of the ViewWidget. The copy to clipboad method allows each variant to be quickly accessed and pasted into their wiki.

To facilitate this a new macro wikitext-example-compact has being added to 
$:/editions/tw5.com/wikitext-macros

* Update ViewWidget.tid

Two colons added to maintain indenting in new content.