Remove the x-requested-with header from the tm-http-request message
Extraneous headers can interfere with CORS validation when hitting external APIs, due to tightly configured `Access-Control-Allow-Headers` headers.  Until there's a way to disable these headers, it's probably better (i.e. most flexible) to expect users to set them deliberately.  

In this case, I ran into an API that uses cors and the above header, and was unable to make a request to it without making the changes in this PR.

I'm also open to modifying the message to allow disabling these headers, but this was the simplest solution I could come up with.