Add TiddlyPWA to GettingStarted
Apologies @valpackett I had not realised that TiddlyPWA was not listed.

I'd welcome any improvements to the listing. I copied the functional description from your website but might be a good opportunity to sell the benefits more.

I'm doing this as part of a review and improvement of the GettingStarted tiddler.